Refactor alert validator functions to helper function
