+++ app/test/e2e/email-notifications/glad-alerts-emails.spec.js
@@ -6,7 +6,6 @@ const redis = require('redis');
 
 const Subscription = require('models/subscription');
 const Statistic = require('models/statistic');
-const AlertUrlService = require('services/alertUrlService');
 const AlertQueue = require('queues/alert.queue');
 const EmailHelpersService = require('services/emailHelpersService');
 
@@ -15,7 +14,11 @@ const { createSubscription } = require('../utils/helpers');
 const { createMockAlertsQuery, createMockGeostore } = require('../utils/mock');
 const { ROLES } = require('../utils/test.constants');
 
-const { assertSubscriptionStats, bootstrapEmailNotificationTests } = require('../utils/helpers/email-notifications');
+const {
+    assertSubscriptionStatsNotificationEvent,
+    bootstrapEmailNotificationTests,
+    validateGLADAlert,
+} = require('../utils/helpers/email-notifications');
 
 nock.disableNetConnect();
 nock.enableNetConnect(process.env.HOST_IP);
@@ -26,77 +29,6 @@ chai.use(require('chai-datetime'));
 const CHANNEL = config.get('apiGateway.queueName');
 const redisClient = redis.createClient({ url: config.get('redis.url') });
 redisClient.subscribe(CHANNEL);
-
-const validateGLADAlert = (
-    jsonMessage,
-    beginDate,
-    endDate,
-    sub,
-    lang = 'en',
-    gladFrequency = 'average',
-    areaName = 'Custom Area',
-    geostoreId = '423e5dfb0448e692f97b590c61f45f22',
-) => {
-    jsonMessage.should.have.property('sender').and.equal('gfw');
-    jsonMessage.should.have.property('data').and.be.a('object');
-
-    jsonMessage.should.have.property('recipients').and.be.a('array').and.length(1);
-    jsonMessage.recipients[0].should.be.an('object')
-        .and.have.property('address')
-        .and.have.property('email')
-        .and.equal('subscription-recipient@vizzuality.com');
-    jsonMessage.data.should.have.property('glad_frequency').and.equal(gladFrequency);
-    jsonMessage.data.should.have.property('month').and.equal(beginDate.format('MMMM'));
-    jsonMessage.data.should.have.property('year').and.equal(beginDate.format('YYYY'));
-    jsonMessage.data.should.have.property('week_of').and.equal(`${beginDate.format('DD MMM')}`);
-    jsonMessage.data.should.have.property('week_start').and.equal(beginDate.format('DD/MM/YYYY'));
-    jsonMessage.data.should.have.property('week_end').and.equal(endDate.format('DD/MM/YYYY'));
-    jsonMessage.data.should.have.property('priority_areas').and.deep.equal({
-        intact_forest: 6,
-        primary_forest: 7,
-        peat: 8,
-        protected_areas: 9,
-        plantations: 10,
-        other: 11
-    });
-    jsonMessage.data.should.have.property('glad_count').and.equal(51);
-    jsonMessage.data.should.have.property('alerts').and.have.length(6).and.deep.equal([
-        { alert_type: 'GLAD', date: '10/10/2019 00:10 UTC' },
-        { alert_type: 'GLAD', date: '11/10/2019 00:10 UTC' },
-        { alert_type: 'GLAD', date: '12/10/2019 00:10 UTC' },
-        { alert_type: 'GLAD', date: '13/10/2019 00:10 UTC' },
-        { alert_type: 'GLAD', date: '14/10/2019 00:10 UTC' },
-        { alert_type: 'GLAD', date: '15/10/2019 00:10 UTC' },
-    ]);
-
-    // Keeping this for backwards compatibility
-    jsonMessage.data.should.have.property('alert_count').and.equal(51);
-    jsonMessage.data.should.have.property('alert_date_begin').and.equal(moment(beginDate).format('YYYY-MM-DD'));
-    jsonMessage.data.should.have.property('alert_date_end').and.equal(moment(endDate).format('YYYY-MM-DD'));
-    jsonMessage.data.should.have.property('alert_link').and.equal(AlertUrlService.generate(
-        sub,
-        {
-            name: 'umd_as_it_happens',
-            slug: 'glad-alerts',
-            subscription: true,
-            datasetId: 'bfd1d211-8106-4393-86c3-9e1ab2ee1b9b',
-            layerId: '8e4a527d-1bcd-4a12-82b0-5a108ffec452'
-        },
-        beginDate,
-        endDate,
-    ));
-    jsonMessage.data.should.have.property('alert_name').and.equal(sub.name);
-    jsonMessage.data.should.have.property('layerSlug').and.equal('glad-alerts');
-    jsonMessage.data.should.have.property('selected_area').and.equal(areaName);
-    jsonMessage.data.should.have.property('subscriptions_url').and.equal(`http://staging.globalforestwatch.org/my-gfw?lang=${lang}`);
-    jsonMessage.data.should.have.property('unsubscribe_url').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/subscriptions/${sub.id}/unsubscribe?redirect=true&lang=${lang}`);
-    jsonMessage.data.should.have.property('downloadUrls');
-    // eslint-disable-next-line max-len
-    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=${geostoreId}&format=csv`);
-    // eslint-disable-next-line max-len
-    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=${geostoreId}&format=json`);
-    jsonMessage.data.should.have.property('value').and.equal(51);
-};
 
 describe('GLAD alert emails', () => {
 
@@ -133,7 +65,7 @@ describe('GLAD alert emails', () => {
                     validateGLADAlert(jsonMessage, beginDate, endDate, subscriptionOne);
                     break;
                 case 'subscriptions-stats':
-                    assertSubscriptionStats(jsonMessage, subscriptionOne);
+                    assertSubscriptionStatsNotificationEvent(jsonMessage, subscriptionOne);
                     break;
                 default:
                     should.fail('Unsupported message type: ', jsonMessage.template);
@@ -172,7 +104,7 @@ describe('GLAD alert emails', () => {
                     validateGLADAlert(jsonMessage, beginDate, endDate, subscriptionOne, 'fr', 'moyenne');
                     break;
                 case 'subscriptions-stats':
-                    assertSubscriptionStats(jsonMessage, subscriptionOne);
+                    assertSubscriptionStatsNotificationEvent(jsonMessage, subscriptionOne);
                     break;
                 default:
                     should.fail('Unsupported message type: ', jsonMessage.template);
@@ -211,7 +143,7 @@ describe('GLAD alert emails', () => {
                     validateGLADAlert(jsonMessage, beginDate, endDate, subscriptionOne, 'zh', '平均');
                     break;
                 case 'subscriptions-stats':
-                    assertSubscriptionStats(jsonMessage, subscriptionOne);
+                    assertSubscriptionStatsNotificationEvent(jsonMessage, subscriptionOne);
                     break;
                 default:
                     should.fail('Unsupported message type: ', jsonMessage.template);
@@ -260,7 +192,7 @@ describe('GLAD alert emails', () => {
                     );
                     break;
                 case 'subscriptions-stats':
-                    assertSubscriptionStats(jsonMessage, subscriptionOne);
+                    assertSubscriptionStatsNotificationEvent(jsonMessage, subscriptionOne);
                     break;
                 default:
                     should.fail('Unsupported message type: ', jsonMessage.template);
@@ -309,7 +241,7 @@ describe('GLAD alert emails', () => {
                     );
                     break;
                 case 'subscriptions-stats':
-                    assertSubscriptionStats(jsonMessage, subscriptionOne);
+                    assertSubscriptionStatsNotificationEvent(jsonMessage, subscriptionOne);
                     break;
                 default:
                     should.fail('Unsupported message type: ', jsonMessage.template);
@@ -358,7 +290,7 @@ describe('GLAD alert emails', () => {
                     );
                     break;
                 case 'subscriptions-stats':
-                    assertSubscriptionStats(jsonMessage, subscriptionOne);
+                    assertSubscriptionStatsNotificationEvent(jsonMessage, subscriptionOne);
                     break;
                 default:
                     should.fail('Unsupported message type: ', jsonMessage.template);
@@ -407,7 +339,7 @@ describe('GLAD alert emails', () => {
                     );
                     break;
                 case 'subscriptions-stats':
-                    assertSubscriptionStats(jsonMessage, subscriptionOne);
+                    assertSubscriptionStatsNotificationEvent(jsonMessage, subscriptionOne);
                     break;
                 default:
                     should.fail('Unsupported message type: ', jsonMessage.template);
@@ -456,7 +388,7 @@ describe('GLAD alert emails', () => {
                     );
                     break;
                 case 'subscriptions-stats':
-                    assertSubscriptionStats(jsonMessage, subscriptionOne);
+                    assertSubscriptionStatsNotificationEvent(jsonMessage, subscriptionOne);
                     break;
                 default:
                     should.fail('Unsupported message type: ', jsonMessage.template);
